Moved views to internal.py
