+++ corehq/apps/domain/views/__init__.py
@@ -42,7 +42,6 @@ from corehq.apps.domain.views.fixtures import (
     CalendarFixtureConfigView,
     LocationFixtureConfigView,
 )
-'''
 from corehq.apps.domain.views.internal import (
     calculated_properties,
     toggle_diff,
@@ -54,7 +53,6 @@ from corehq.apps.domain.views.internal import (
     EditInternalDomainInfoView,
     EditInternalCalculationsView,
 )
-'''
 from corehq.apps.domain.views.pro_bono import (
     ProBonoMixin,
     ProBonoStaticView,
